Complete rewrite of the service
The original task was to get a more direct response from the /update
endpoint when a change occurs. The problem is that the Delta service
waits for each callback to finish before returning a result on an
update. This is fine but this service can potentially takes a very long
time to proceed some requests.

For instance, if you add a new repository and ask the API to "Compose
Up" the pipeline, it will first pull all the images, build all the
images that need to be build, then only start. If the frontend needs to
wait for all of these operations to be done to continue, it will become
a sluggish UI to use.

Now all the operations are started asynchronously with subprocesses.
When you ask for a "Compose Up", the service will return after writing
a status "Starting" to the database but a thread and a subprocess for
the actual action will be running in background (this includes a timeout
protection).

In this commit, all the endpoints have also been removed. The idea from
the beginning is that the only available endpoint is /update for the
Delta service. The frontend should communicate its changes only by
changing data in the database. This is almost the case since there is
still one endpoint that I couldn't get rid off: the one to get the
service logs.