Removes duplicated calls
Now that the containers are monitored, the status of the services is
updated twice: first because of the Docker event and a second time
because the docker-compose project actually finish.

This change will keep the update of the status only where it makes
sense (Up instead of Started for example).