Do not run updates in background tasks
This is normally not recommended since the service must return as soon
as possible from the callback but this behavior actually duplicated with
the ActionScheduler.

The ActionScheduler already run the tasks in background and in serie to
avoid multiple pipeline actions to be run at the same time.

By removing the upper tasks we resolve 2 problems:

1.  Any error that could happen before an ActionScheduler is registered
    will be showed properly in the application logs.
2.  When writing tests, at the end of an update call to the SPARQL
    endpoint, you can be sure there will be an executer for the
    ActionScheduler running.

The unfortunate consequence is that the query calls to the database made
during the treatment of the request must not exceeds 30 seconds.