Fixed ActionScheduler storage of executers
Executers are stored in a class variable. Unfortunately, the code was
using always the same class to look for its executer instead of
dynamically selecting the class of the current instance.

The second issue was that an executer can be cancelled by raising a
specific exception during an action. If this exception is raised while
the scheduler is already canceling (method cancel() being executed), the
cancellation will fail to find the executer in the list at the end
because it has been deleted already.