Enqueue the container events to the ActionScheduler
This will prevent the actions to happen at the same time: when you turn
a pipeline up, the state of the pipeline will stay unchanged until it
reaches the end of the process. Then all the events of Docker will
apply: the scaling value of the services will be updated, the status of
the services and also (unfortunately) the state of the pipeline.

This change will allow integration tests to work on the whole process.