Fixed: better application cleanup & tests
If the database is not accessible, it will fail to start.

If Docker is not accessible, it will fail to start.

If an error occur during the monitoring of the Docker event that is not
an error in the event processing itself, it will quit the application.

The error messages in those cases has been reduced (no traceback).

If an unexpected error not related to connection error during startup
occur, it will display a traceback and exit the application.